pyproject.toml
[project]
name = "exile-ai-test-platform-server"
version = "0.1.0"
description = "exile ai test platform server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.10.4",
    "httpx>=0.27.2",
    "bcrypt>=4.2.0",
    "celery>=5.4.0",
    "fastapi>=0.115.0",
    "loguru>=0.7.2",
    "pydantic-settings>=2.5.2",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2024.1",
    "redis>=5.0.8",
    "shortuuid>=1.0.13",
    "sqlalchemy>=2.0.36",
    "asyncmy>=0.2.9",
    "asyncpg>=0.30.0",
    "uvicorn[standard]>=0.30.6",
    "email-validator>=2.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.60",
    "langchain-community>=0.4.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.9",
]

[dependency-groups]
dev = [
    "alembic>=1.14.1",
    "pytest>=8.3.2",
]

[tool.uv]
package = false
